- edit (trim) video button - renaming files
