+++ app/src/main/java/org/mozilla/tiktokreporter/studyonboarding/StudyOnboardingScreenViewModel.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import org.mozilla.tiktokreporter.data.model.Form
 import org.mozilla.tiktokreporter.data.model.OnboardingStep
-import org.mozilla.tiktokreporter.repository.TikTokReporterRepository
+import org.mozilla.tiktokreporter.TikTokReporterRepository
 import org.mozilla.tiktokreporter.util.OneTimeEvent
 import org.mozilla.tiktokreporter.util.toOneTimeEvent
 import javax.inject.Inject
